Extract ListBox elements to own components
