+++ ui/src/uiComponents/Listbox.spec.tsx
@@ -1,5 +1,4 @@
 import { fireEvent, screen } from '@testing-library/react';
-import React from 'react';
 import { render } from '../utils/test-utils';
 import { Listbox } from './Listbox';
