+++ ui/src/uiComponents/index.ts
@@ -11,6 +11,8 @@ export * from './FormContainer';
 export * from './IconButton';
 export * from './IconToggle';
 export * from './Listbox';
+export * from './ListboxButton';
+export * from './ListboxOptions';
 export * from './LoadingOverlay';
 export * from './LocatorButton';
 export * from './Pagination';
